# Refs #28459 -- Improved performance of ValuesIterable.